+++ cube/base/ogldev_math_3d.h
@@ -22,6 +22,7 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
+#include <math.h>
 #ifdef WIN32
 #define _USE_MATH_DEFINES 
 #include <cmath>
@@ -35,16 +36,9 @@
 #define ZERO_MEM(a) memset(a, 0, sizeof(a))
 #define ARRAY_SIZE_IN_ELEMENTS(a) (sizeof(a)/sizeof(a[0]))
 
-#ifdef WIN32
-#define SNPRINTF _snprintf_s
+#define SNPRINTF snprintf
 #define RANDOM rand
 #define SRANDOM srand((unsigned)time(NULL))
-float fmax(float a, float b);
-#else
-#define SNPRINTF snprintf
-#define RANDOM random
-#define SRANDOM srandom(getpid())
-#endif
 
 #define INVALID_UNIFORM_LOCATION 0xffffffff
 #define INVALID_OGL_VALUE 0xffffffff