+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 here is the changelog of the CubeEngine
 
+###alpha v0.11b 3.24.2015###
+* fix shader compile error in some machine.
+* fix engine C++ part compile error in some machine.
+
 ###alpha v0.11 3.23.2015###
 * add two water simulation classes - waterNaive & waterPG(Projected grid water),use waterNaive to simulate a small water area, use waterPG to simulate unlimited water surface like ocean.
 * add RenderTarget class, you can use RenderTarget to control the render pipeline.for example you can use it to simulate mirror.